Features on this app
1. As a user at the search screen,
a. When I enter a name of a movie (e.g. "Batman", "Rocky") in the search box and
tap on "search button"
b. Then I should see a new list view with the following rows
i. Movie Poster
ii. Movie name
iii. Release date
iv. Full Movie Overview
2. As a user at the Movie List Screen,
a. When I scroll to the bottom of list
b. Then next page should load if available
3. As a user at the search screen,
a. When I enter a name of a movie that doesn’t exist in the search box and tap on
"search button",
b. Then, an alert box should appear and display an error message.
c. All types of error should be handled
4. As a user at the search screen,
a. When I tap and focus into the search box
b. Then an auto suggest list view will display below the search box showing my last
10 successful queries (exclude suggestions that return errors)
c. Suggestions should be persisted.
5. As a user at the search screen with the auto suggest list view shown,
a. When I select a suggestion
b. Then the search results of the suggestion will be shown.